pyproject.toml
[project]
name = "manager"
version = "0.1.0"
description = "File manager compiler"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "antlr4-python3-runtime>=4.13.2",
    "cyclopts>=3.13.1",
]

[project.scripts]
manager = "manager.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
